requirements.txt
# =============================================================================
# Odisha Flood Validation System - Requirements
# =============================================================================
# Python 3.10+ required
# Install with: pip install -r requirements.txt
#
# Note: GDAL should be installed via conda for better compatibility:
#   conda install -c conda-forge gdal=3.7.1
# =============================================================================

# -----------------------------------------------------------------------------
# Core Data Science Stack
# -----------------------------------------------------------------------------
pandas==2.1.0
numpy==1.24.3
scipy==1.11.1

# -----------------------------------------------------------------------------
# Machine Learning
# -----------------------------------------------------------------------------
scikit-learn==1.3.0
joblib==1.3.2

# -----------------------------------------------------------------------------
# Geospatial Processing
# -----------------------------------------------------------------------------
# Note: Install GDAL via conda first: conda install -c conda-forge gdal=3.7.1
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
fiona==1.9.4
pyproj==3.6.0

# For coordinate transformations
geopy==2.4.0

# Note: WhiteboxTools installed separately from:
# https://www.whiteboxgeo.com/download-whiteboxtools/

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
geoalchemy2==0.14.1

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6

# CORS and security
python-jose[cryptography]==3.3.0

# -----------------------------------------------------------------------------
# HTTP Client & API Integration
# -----------------------------------------------------------------------------
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5

# Twitter API
tweepy==4.14.0

# Rate limiting
ratelimit==2.2.1

# -----------------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------------
matplotlib==3.7.2
seaborn==0.12.2

# Interactive notebooks
jupyterlab==4.0.5
ipywidgets==8.1.0

# Maps in notebooks
folium==0.14.0
mapclassify==2.6.0

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
pyyaml==6.0.1

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.24.1  # For testing FastAPI

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1

# Type stubs
pandas-stubs==2.0.3.230814
types-requests==2.31.0.2

# -----------------------------------------------------------------------------
# Documentation
# -----------------------------------------------------------------------------
mkdocs==1.5.2
mkdocs-material==9.2.5

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
tqdm==4.66.1      # Progress bars
click==8.1.7      # CLI tools
rich==13.5.2      # Rich terminal output
loguru==0.7.0     # Better logging

# Date/time handling
python-dateutil==2.8.2

# -----------------------------------------------------------------------------
# Optional: Real-time Processing
# -----------------------------------------------------------------------------
# Uncomment if using Kafka for streaming
# kafka-python==2.0.2
# confluent-kafka==2.2.0

# -----------------------------------------------------------------------------
# Optional: Cloud Integration
# -----------------------------------------------------------------------------
# Uncomment for Google Cloud
# google-cloud-storage==2.10.0
# earthengine-api==0.1.365
